fix(server): relocate boot guards to buildApp + .min(1) on App secrets (codex P1-8)
Two coupled fixes to close the "blank webhookSecret = forgeable
HMAC" hole codex flagged:

1. Boot-time config-shape validation moves from
   `packages/server/src/index.ts` to a new `boot-guards.ts` module
   invoked from `buildApp`. Both server entry points now run the guard:
     - the standalone bin (`packages/server/src/index.ts`)
     - the CLI server-start path
       (`packages/command/src/core/server.ts` → buildApp)
   Previously the CLI path skipped the guard entirely — a misconfigured
   `first-tree-hub server start` could boot with `webhookSecret=""` and
   `createHmac("sha256", "")` would happily verify any attacker-computed
   SHA-256 of the request body.

2. `.min(1)` added to every `oauth.githubApp.*` field in the shared
   config schema. Blank env values (e.g. `FIRST_TREE_HUB_GITHUB_APP_WEBHOOK_SECRET=`)
   now fail Zod parse at config-init time, before they can reach the
   boot guard or — worse — the webhook handler. Same defense applies
   uniformly to all five fields: blank appId would make App JWTs
   unverifiable upstream, blank clientId would let GitHub round-trip
   an anonymous OAuth, etc.

Belt-and-braces: the boot guard ALSO trims and checks for the
all-empty case (defense in depth in case the schema is ever loosened
or a future field is added without the `.min(1)`), and verifies the
PEM looks like a real PKCS#8 PEM header.

Test helper (`__tests__/helpers.ts`) had to be updated: its stub
`privateKeyPem` was a non-PEM placeholder string that the relocated
guard now correctly rejects. Replaced with a header-only stub. The
test path never actually signs with it; the body is junk.

Verified: 738 / 738 server suite green. Boot guard's "half-configured"
and "blank PEM" checks remain belt-and-braces — the Zod `.min(1)`
catches them first, but the guard fires if someone ever weakens the
schema.
